+++ src/app/(main)/layout.tsx
@@ -1,0 +1,9 @@
+import Layout from "@/components/common/layout";
+
+export default function MainLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <Layout>{children}</Layout>;
+}
